What we have so far
We were trying to look at different ways to implement the map into the site; one such option was to split the js code into a Map.js file and the indexView.js file. While declaring the Map class in the Map.js file, ESLint doesn't recognize the class and so we can't actually see our results in the browser.

We also realized that we needed an extra library (leaflet, in addition to react-leaflet), and so we tried to do "npm install leaflet --save" in the frontend directory, but we're not sure if this was the right option.

Thanks for your help!