Scaled map on landing page to fit between navbar and footer
This required some edits to the map and the navbar and footer. After realizing that many of the problems of scale came from some of the overlap in display, I rescaled some of the dimensions of the navbar and footer to allow for more responsive design. This way, I could constrain the mapContainer to fit between a navbar and footer with relatively defined margins that try to preserve size proportions as much as possible. (e.g. redefining height and width from px to vh, vw, percentages, rems, etc). At this point, I still need to account for scaling the logos in the footer to make sure that the map and footer can be displayed properly on even mobile devices.