Added photographer data to dropdown; still working on cool dropdown
After an extended period of time, I've managed to get the photographer data to show up in the single-select dropdown. Next steps: Convert this to a multi-select dropdown.

Also, I realized I'm going to have to take a different approach to try and get something that resembles the cool multi-select dropdown, but I have a good idea of where to go from here.