Fixing the broken links on the photographer page
Previous implementation of the photographers page was loading photos that were stored directly in the codebase, however we have a large amount of photos. I updated 'urls.py' to add a file path that returns one photo from a given photographer and wrote the function 'get_photographer_photos' in 'views.py' that executes the function when it's called. 'PhotographerListView.js' was updated to use this file path for each photographer being loaded on the page.